Java Compiler Warnings changed to be more strict. Code cleanup: unused code removed. initial config.ini modified to have the correct escaping.
Reviewed-on: https://git.eclipse.org/r/17099
Tested-by: Hudson CI
IP-Clean: Matthias Villiger <mvi@bsiag.com>

+++ org.eclipse.scout.sdk.testing/src/org/eclipse/scout/sdk/testing/SdkAssert.java
@@ -641,7 +641,7 @@ public class SdkAssert extends Assert {
       fail(message);
     }
 
-    assertEquals(message, (Double) memberOrderNr, (Double) orderNr);
+    assertEquals(message, memberOrderNr, orderNr);
   }
 
   /**
